Update Push-Relabel (5).cpp

+++ Implementations/12 - Graphs Hard (4)/12.3 - Flows/Push-Relabel (5).cpp
@@ -16,7 +16,7 @@ template <int SZ> struct PushRelabel {
     bool active[SZ];
     vi B[SZ];
     
-    void addEdge(int u, int v, int C) {
+    void addEdge(int u, int v, ll C) {
         Edge a{v, 0, C, sz(adj[v])};
         Edge b{u, 0, 0, sz(adj[u])};
         adj[u].pb(a), adj[v].pb(b);
